% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesDiffIRT-package.R
\docType{package}
\name{BayesDiffIRT-package}
\alias{BayesDiffIRT}
\alias{BayesDiffIRT-package}
\title{BayesDiffIRT: Bayesian Sampling of Diffusion Item Response Theory Models}
\description{
Tools for simulating, fitting, and analyzing diffusion IRT models for
responses and response times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ManuelRausch/BayesDiffIRT}
  \item Report bugs at \url{https://github.com/ManuelRausch/BayesDiffIRT/issues}
}

}
\author{
\strong{Maintainer}: Manuel Rausch \email{manuel.rausch@aau.at} (\href{https://orcid.org/0000-0002-5805-5544}{ORCID})

Other contributors:
\itemize{
  \item Rainer Alexandrowicz [contributor]
}

}
\keyword{internal}
