% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bayesDiffIRTexample}
\alias{bayesDiffIRTexample}
\title{Example Data for \code{BayesDiffIRT}}
\format{
A data frame with 200 rows and 4 variables:
\describe{
\item{person}{Integer person identifier.}
\item{item}{Integer item identifier.}
\item{response}{Binary response indicator. Values are 0 and 1.}
\item{rt}{Response time in seconds.}
}
}
\usage{
bayesDiffIRTexample
}
\description{
A small simulated data set for demonstrating the interface of
\code{BayesDiffIRT}. The data contain binary responses and response times
for multiple persons and items.
}
\details{
The data are simulated and are intended only for examples, tests, and
demonstrations.
}
\examples{
data("bayesDiffIRTexample")
head(bayesDiffIRTexample)
str(bayesDiffIRTexample)
}
\keyword{datasets}
