% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3extract.R
\name{extractSamples}
\alias{extractSamples}
\alias{extractSamples.BayesDiffIRTfit}
\title{Extract posterior samples}
\usage{
extractSamples(object, parameters = NULL, ...)

\method{extractSamples}{BayesDiffIRTfit}(object, parameters = NULL, ...)
}
\arguments{
\item{object}{A fitted model object of class \code{BayesDiffIRTfit}.}

\item{parameters}{Optional character vector specifying which parameters to
extract. If \code{NULL}, samples for all parameters are returned. Parameter
names can refer to scalar parameters, vector parameters, or individual
indexed parameters, depending on how they are stored in the fitted Stan
model. For example, \code{"theta"} extracts all person-specific
\code{theta} parameters, whereas \code{"theta[1]"} extracts only the first
one.}

\item{...}{Currently unused.}
}
\value{
A \code{draws_df} object containing posterior samples.
}
\description{
Extract posterior draws from a fitted \code{BayesDiffIRT} model.
}
\examples{
\dontrun{
samples <- extractSamples(fit)
theta.samples <- extractSamples(fit, parameters = "theta")
selected.samples <- extractSamples(fit, parameters = c("theta", "gamma", "a"))
}

}
