% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3PlotItemSurface.R
\name{plotResponseSurface}
\alias{plotResponseSurface}
\alias{plotResponseSurface.BayesDiffIRTfit}
\title{Plot response-probability surfaces}
\usage{
plotResponseSurface(object, ...)

\method{plotResponseSurface}{BayesDiffIRTfit}(
  object,
  item,
  theta.range = NULL,
  gamma.range = NULL,
  grid.size = 60,
  contours = c(0.1, 0.3, 0.5, 0.7, 0.9),
  ndraws = 200
)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to methods.}

\item{item}{Integer. Indexes the item for which the response-probability
surface should be plotted.}

\item{grid.size}{Integer. Number of grid points used for each axis. The total
number of evaluated grid points is grid.size * grid.size.
Defaults to 60.}

\item{contours}{Numeric vector. Response-probability values at which contour
lines should be drawn. Defaults to c(.1, .3, .5, .7, .9).}

\item{ndraws}{Integer. Number of posterior draws used to average the response
probability at each grid point. Defaults to 200.}

\item{theta.range.}{Numeric vector of length two. Range of theta
values shown on the x-axis. If NULL, the range is set to the 1st and 99th
percentiles of the posterior draws of thetas form all items.}

\item{gamma.range.}{Numeric vector of length two. Range of gamma
values shown on the y-axis. If NULL, the range is set to the 1st and 99th
percentiles of the posterior draws of gamma from all items,}
}
\value{
A ggplot2 object.

a ggplot object
}
\description{
\code{plotResponseSurface()} plots the posterior mean response probability for a
selected item over a two-dimensional grid of person parameters. The x-axis
represents the latent trait parameter theta, and the y-axis represents the
response caution parameter gamma. For each grid point, the response probability
is averaged over a random subset of posterior draws of the selected item's parameters.
The resulting surface is shown as a raster plot with contour lines.
}
\details{
The method computes the predicted probability of a correct response (for ability
tests or of item acceptance (for survey questions) for a selected item across a grid of theta and gamma values.
The response probabilities are calculated using code{\link[rtdists:Diffusion]{rdiffusion}}\insertCite{Singmann2026}{BayesDiffIRT}..
}
\examples{
\dontrun{
data("bayesDiffIRTexample")
fit <- fitBayesDiffIRT(data=bayesDiffIRTexample, model = "d")

plotResponseSurface(fit, item = 1)

plotResponseSurface(
fit,
item = 2,
theta.range = c(-3, 3),
gamma.range = c(0.5, 3),
grid.size = 80,
contours = c(.25, .5, .75),
ndraws = 500
)
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{fitBayesDiffIRT}},
\code{\link[rtdists:Diffusion]{rdiffusion}}
}
\author{
Manuel Rausch, \email{manuel.rausch@aau.at}
}
