% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Prior Definitions for \strong{BayesDiffIRTfit} models}
\usage{
prior(dist, class, coef = NULL)
}
\arguments{
\item{dist}{An expression defining a distribution in \strong{Stan} language (see examples)}

\item{class}{\code{character} The parameter class. The following parameter classes are available:
\itemize{
\item \code{omega_theta} standard deviation of \theta, the latent trait.
\item \code{omega_gamma} standard deviation of \gamma, the response caution parameter.
\item \code{nu} item difficulty,
\item \code{a} item time pressure,
\item \code{tnd} person-specific non-decision time.
}}

\item{description}{coef currently not implemented.}
}
\description{
Define priors for specific parameters or classes of parameters.
}
