% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3diagnostics.R
\name{checkDiagnostics}
\alias{checkDiagnostics}
\title{Run diagnostic checks for BayesDiffIRTfit objects.}
\usage{
checkDiagnostics(object, ...)
}
\arguments{
\item{object}{A fitted BayesDiffIRTfit object.}

\item{...}{Further arguments passed to methods.}
}
\description{
Run diagnostic checks for BayesDiffIRTfit objects.
}
