% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBayesDiffIRT.R
\name{fitBayesDiffIRT}
\alias{fitBayesDiffIRT}
\title{Fit Bayesian diffusion item-response theory models}
\usage{
fitBayesDiffIRT(
  data,
  rt = "rt",
  resp = "resp",
  sbj = "sbj",
  item = "item",
  model = "d",
  priors = NULL,
  seed = NULL,
  n.chains = 4,
  n.cores = n.chains,
  n.warmup = 1000,
  n.samples = 1000,
  adapt.delta = 0.95,
  max.treedepth = 12,
  init = NULL,
  refresh = 200,
  diagnostic.warnings = TRUE,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{data}{a \code{data.frame} with different row indicating responses to different items.
It must include at least three columns:
\itemize{
\item One column must provide response times. They should be numeric and measured in seconds.
\item One column must provide integer responses.
\itemize{
\item For ability tests, the value should be 0 for incorrect responses
or correct responses.
\item For questionnaires, the value should be 0 for rejected items and 1 for accepted items.
}
}}

\item{rt}{\code{character} the name of the column in data that provides reaction times.}

\item{model}{\code{character} the name of the model of which the parameters should be sampled. So far,
The following models have been implemented:  'D', 'Q'}

\item{priors}{BayesDiffIRTPrior object, or a list of BayesDiffIRTPrior objects, created with prior().}

\item{seed}{A seed for the random number generators to be passed to Stan.}

\item{n.chains}{Number of Markov chains.}

\item{n.cores}{Number of chains to run in parallel.}

\item{n.warmup}{Number of warmup iterations per chain.}

\item{n.samples}{Number of post-warmup sampling iterations per chain.}

\item{...}{Additional arguments passed to \code{cmdstanr::CmdStanModel$sample()}.
This is intended for advanced users. Common sampling controls such as
\code{chains}, \code{parallel_chains}, \code{iter_warmup}, \code{iter_sampling},
\code{adapt_delta}, \code{max_treedepth}, and \code{init} should be supplied via the
corresponding arguments of \code{fitBayesDiffIRT()}.}

\item{response}{\code{character} the name of the column in data that provides reaction times.}

\item{id}{\code{character} the name of the column in data in which responses are found.}
}
\value{
An object of class \code{BayesDiffIRTfit}. Available methods include \code{print}, \code{summary}, \code{plot} and \code{checkDiagnostics}.
}
\description{
\code{fitBayesDiffIRT} samples item and subject parameters
from the posterior distributions for diffusion item response theory
models. For this purpose, NUTS sampling as implemented in STAN is used. See Details for the mathematical specification of the implemented models and
their parameters.
}
\details{
\code{BayesDiffIRT} samples from the posterior distributions of item and subject parameters
and from posterior predictive distributions of responses and reaction times for
diffusion item response theory models
\insertCite{van_der_maas_cognitive_2011,molenaar_fitting_2015,tuerlinckx_two_2005}{BayesDiffIRT}.
\subsection{Mathematical description of models}{

According to the drift diffusion decision model, in each moment, the sensory
system generates new sensory evidence about which of the two possible choice options
is the correct one. This momentary evidence is drawn from Gaussian
distributions and accumulated over time, i.e. the newly
acquired evidence is constantly added to the evidence collected up
to that moment. This accumulation process is bounded by an upper
}
}
\examples{
print("Coming!")
}
\references{
\insertAllCited{}
}
\author{
Manuel Rausch, \email{manuel.rausch@aau.at}
}
