% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitBayesDiffIRT.R
\name{fitBayesDiffIRT}
\alias{fitBayesDiffIRT}
\title{Fit Bayesian diffusion item-response theory models}
\usage{
fitBayesDiffIRT(
  data,
  rt = "rt",
  resp = "resp",
  sbj = "sbj",
  item = "item",
  model = "d",
  prior = NULL,
  seed = 42,
  chains = 3,
  parallel_chains = 1,
  iter_warmup = 1000,
  iter_sampling = 2000
)
}
\arguments{
\item{data}{a \code{data.frame} with different row indicating responses to different items.
It must include at least three columns:
\itemize{
\item One column must provide response times. They should be numeric and measured in seconds.
\item One column must provide integer responses.
\itemize{
\item For ability tests, the value should be 0 for incorrect responses
or correct responses.
\item For questionnaires, the value should be 0 for rejected items and 1 for accepted items.
}
}}

\item{rt}{\code{character} the name of the column in data that provides reaction times.}

\item{model}{\code{character} the name of the model of which the parameters should be sampled. So far,
the following models have been implemented:  Models implemented so far: 'D', 'Q'}

\item{prior}{BayesDiffIRTPrior object, or a list of BayesDiffIRTPrior objects, created with prior().}

\item{seed}{A seed for the random number generators to pass to Stan.}

\item{chains}{\code{integer} the number of Markov chains to be run.}

\item{parallel_chains}{\code{integer} The maximum number of MCMC chains to run in parallel.}

\item{iter_warmup}{\code{integer} number of warmup (aka burnin) iterations.}

\item{iter_sampling}{`integer The number of post-warmup iterations to run per chain.}

\item{response}{\code{character} the name of the column in data that provides reaction times.}

\item{id}{\code{character} the name of the column in data in which responses are foundedn.}
}
\value{
An object of class \code{BayesDiffIRTfit}. Available methods include \code{print}, \code{summary}, and \code{plot}.
}
\description{
\code{fitBayesDiffIR} samples item and subject parameters
from the posterior distributions for diffusion item response theory
models. For this purpose, NUTS sampling is used. See Details for the mathematical specification of the implemented models and
their parameters.
}
\details{
\code{BayesDiffIRT} samples from the posterior distributions of item and subject parameters
and from posterior predictive distributions of responses and reaction times for
diffusion item response theory models (REFERENCE MISSING).
\subsection{Mathematical description of models}{

According to the drift diffusion decision model, in each moment, the sensory
system generates new sensory evidence about which of the two possible choice options
is the correct one. This momentary evidence is drawn from Gaussian
distributions and accumulated over time, i.e. the newly
acquired evidence is constantly added to the evidence collected up
to that moment. This accumulation process is bounded by an upper
}
}
\examples{
# 1) Prepare dataset
data(extraversion, package="diffIRT")
x=extraversion[,1:10]
x <- cbind(1:nrow(x), x)
colnames(x) <- c("sbj", paste0("Item", 1:10))
x <- tidyr::pivot_longer(as.data.frame(x), cols=Item1:Item10,
                         names_to="item",
                         values_to = "resp")

rt=extraversion[,11:20]
rt <-  cbind(1:nrow(rt), rt)
colnames(rt) <- c("sbj", paste0("Item", 1:10))
rt <- tidyr::pivot_longer(as.data.frame(rt), cols=Item1:Item10,
                          names_to="item",
                          values_to = "rt")
Extra <- merge(x, rt)

# 2) sample from the posterior
samples <-
  fitBayesDiffIRT(Extra,
                  rt = "rt", resp = "resp", sbj = "sbj",
                  item = "item", model = "d",
                  prior = NULL,
                  seed = 42, chains = 1,
                  parallel_chains = 1,
                  iter_warmup =  1000,
                  iter_sampling = 1000)
summary(samples)
}
\references{
Alexandrowicz, R. W. (2020). The diffusion model visualizer: An interactive tool to understand the diffusion model parameters. Psychological Research, 84(4), 1157-1165. https://doi.org/10.1007/s00426-018-1112-6

Molenaar, D., Tuerlinckx, F., & Maas, H. L. J. V. D. (2015). Fitting Diffusion Item Response Theory Models for Responses and Response Times Using the R Package diffIRT. Journal of Statistical Software, 66(4). https://doi.org/10.18637/jss.v066.i04

Van Der Maas, H. L. J., Molenaar, D., Maris, G., Kievit, R. A., & Borsboom, D. (2011). Cognitive psychology meets psychometric theory: On the relation between process models for decision making and latent variable models for individual differences. Psychological Review, 118(2), 339-356. https://doi.org/10.1037/a0022749
}
\author{
Manuel Rausch, \email{manuel.rausch@aau.at}
}
