% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3ppCheck.R
\name{ppCheck}
\alias{ppCheck}
\alias{ppCheck.BayesDiffIRTfit}
\title{Posterior predictive checks}
\usage{
ppCheck(object, ...)

\method{ppCheck}{BayesDiffIRTfit}(
  object,
  type = c("response"),
  ndraws = 10,
  group = c("none", "item", "person"),
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{A fitted model object or posterior predictive object.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Posterior predictive checks
}
